load_pw_modules: Make it no-op if no pw-cli or pipewire is not running
With this change, the script exits gracefully if ever it runs on a
system where pipewire is not installed or not running.

Note that 'pw-cli quit' was added in pipewire 0.3.38, so we're good
since we require 0.3.58 at least.

While we're here, we also drop a stray comment (it comes from
load_pa_modules.sh, however the code associated with this comment is
gone, cf. [1]), and we replace it by a comment to explain what the next
code block does.

[1]: https://github.com/neutrinolabs/pulseaudio-module-xrdp/blob/791dc3f95cd099adb1075708fb46069e0c1257bb/instfiles/load_pa_modules.sh#L6-L13