Adjust LDFLAGS: drop -module, add PW_LIBS and SPA_LIBS
The "-module" ldflag seems to come from an old Pulseaudio documentation,
authored in 2008, cf. [1] section "Update on July 30 2008". This
documentation suggests to use the following ldflags:

    module_abc_la_LDFLAGS = -module -avoid-version

The ldflag "-avoid-version" is documented at [2] and seems to be right,
thanks to this flag the build produces only a ".so" file, without all
the extra symlinks that are usally associated with libraries.

What about "-module" then? According to the doc [3], there are 2 reasons
to use this particular ldflag:

> libraries meant to be dlopened

I don't think it's relevant here.

> Automake requires that a library’s name start with lib. However, [...]
> you might wish to use a "nonstandard" name. Automake will not complain
> about such nonstandard names if it knows the library being built is a
> libtool module

I think that was the point of this flag. Indeed, Pulseaudio modules used
"nonstandard" names such as "module-pipe-sink.so" instead of
"libmodule-pipe-sink.so". That's probably why the documentation
suggested to use this ldflag.

Back to Pipewire: the modules built are named "libmodule-xxx", that's a
"standard" name, so there's no need to use the "-module" ldflag anymore.

Therefore this commit drops the "-module" ldflag.

Additionally, we add the ldflags of the libs used to build the module:
it's the usual way to do things, and it allows tooling (eg. Debian
packaging tooling) to check that the symbols used by the module are
indeed present in the Pipewire library, and issue a warning otherwise.

For reference, see also the ticket opened upstream [4], that didn't get
much attention ans suggest that the whole thing doesn't matter much.

[1] https://www.freedesktop.org/wiki/Software/PulseAudio/Documentation/Developer/Modules/
[2] https://www.gnu.org/software/libtool/manual/html_node/Link-mode.html
[3] https://www.gnu.org/software/automake/manual/html_node/Libtool-Modules.html
[4] https://gitlab.freedesktop.org/pipewire/pipewire/-/issues/3913